repair: Move repair_sync_boundary to a separate header.
